requirements.txt
Django>=4.0,<5.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
psycopg2>=2.9.0
gunicorn>=20.1.0
pylint>=2.17.5
django-cors-headers>=3.13.0